CU-8678cunf5_Ajustar-e-renomear-camada-Controller_Allan-David-Almeida-de-Albuquerque

+++ src/main/java/com/syst/trades/repository/TradeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.syst.trades.model.Trade;
 
-public interface TradeRepository extends JpaRepository<Trade, Long>{
+public interface TradeRepository extends JpaRepository<Trade, Integer>{
 
 }
